Update to turbo-rails 2 beta and use page morphing
We now broadcast a refresh stream action whenever cards or columns are added, removed, or moved. This will cause Turbo
to fetch the HTML for the page and morph the changes into the DOM while preserving the browser's scroll position.

To avoid an unnecessary refresh for the user who is dragging a card we need to add an X-Turbo-Request-Id header to our
AJAX request and add it to the list of Turbo's recent requests. Turbo will then ignore any refresh stream actions with a
matching request ID.

In order to avoid the add-card form being closed and reset during a page morph caused by another user we have to add a
data-turbo-permanent attribute. However, this prevents those elements from being moved with their columns when they are
re-ordered. To fix this we add an event listener that runs after the page morph and moves any add-card forms that are in
the wrong columns.