Refactor to simplify broadcasting refreshes
Instead of both card and column models broadcasting refreshes, we make both models
`touch` their parent (column in the case of card, and board in the case of column).
This means that changes to a card or column will bubble up and cause the `updated_at`
attribute of the board to be updated. Therfore we can just use a single
`broadcasts_refreshes` on the board model.